クリップのサムネイル取得方式変更
- 画像をロードする上限数の指定(通信量削減)
- 不要なimport削除
- サムネイルの縦横比を1:1に
